class_tasks/task_4/pyproject.toml
[project]
name = "tasks4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = true

[project.scripts]
tasks4 = "tasks4:main"